requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Database
asyncpg==0.29.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# Environment & Config
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AI & NLP
openai==1.10.0

# Web Scraping
beautifulsoup4==4.12.3
lxml==5.1.0
selenium==4.16.0

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring & Logging
sentry-sdk[fastapi]==1.40.0

# Rate Limiting
slowapi==0.1.9

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx-mock==0.15.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
